Discovered and rectified why arrayoutofbounds error occurs
This is due to wrong formatting of the .txt save file. Always ensure
that there are 5 “~~” on each line. The saving function is in place to
ensure this. Adding and deleting tasks from the app is fine. Need to
take extra care when you edit the .txt file yourself, please ensure
proper formatting.